Fix guava filter
